+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Handle tuple types for which the desired default value is `None` (rather than a sequence of `None`s)
 * Make ValidationError inherit from UsageError
 * Fix Defined validations
+* By default do not create an instance of a composite types if none of its parameters are defined
 
 ## 0.2.3 (2019.01.15)